\name{MotorCycles}
\alias{MotorCycles}
\title{Motor Cycles in The Netherlands}
\description{
Time series of stock of motor cyles (two wheels) in The Netherlands
(in thousands).
}
\usage{data("MotorCycles")}

\format{
An annual univariate time series from 1946 to 1993.
}

\source{
Online complements to Franses (1998).

\url{http://www.few.eur.nl/few/people/franses/research/book2.htm}
}

\references{
Franses, P.H. (1998). \emph{Time Series Models for Business and Economic Forecasting}.
Cambridge, UK: Cambridge University Press.
}

\seealso{\code{\link{Franses1998}}}

\examples{
data("MotorCycles")
plot(MotorCycles)
}

\keyword{datasets}
